api/paymentに関する型定義をtypes/api/paymentに移動
