+++ frontend/src/types/api/payment.ts
@@ -1,5 +1,16 @@
 import type { User } from 'types/user'
 
+export type PostPaymentParams = {
+  amount: string
+  detail: string
+  paidAt: string
+}
+
+export type PaymentListGroupByPaidAt = {
+  date: string
+  payments: Payment[]
+}
+
 export type Payment = {
   amount: number
   paidAt: string
