--- frontend/src/types/paymentListGroupByPaidAt.ts
@@ -1,6 +1,0 @@
-import type { Payment } from 'types/payment'
-
-export type PaymentListGroupByPaidAt = {
-  date: string
-  payments: Payment[]
-}
