--- frontend/src/types/postPaymentParams.ts
@@ -1,5 +1,0 @@
-export type PostPaymentParams = {
-  amount: string
-  detail: string
-  paidAt: string
-}
